FCOM-193: moved input/output property classes to Definitions.cs, fixed example values for user CN
