+++ Frends.Community.LDAP/Frends.Community.LDAP.csproj
@@ -44,6 +44,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Definitions.cs" />
     <Compile Include="LdapOperations.cs" />
     <Compile Include="Models\ADFlag.cs" />
     <Compile Include="Models\AdUser.cs" />
